+++ packages/markovian-continuous/test/compile-fail/PairedDifferenceBoundary.hs
@@ -3,4 +3,4 @@ module PairedDifferenceBoundary where
 import Markovian.Continuous.Measure.Exact
 
 badPairedConstructor :: PairedDifferenceReport
-badPairedConstructor = PairedDifferenceReport
+badPairedConstructor = PairedDifferenceReport 0 0 0 0 0 0 0 0 0 EQ undefined